% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{approxNA,funData-method}
\alias{approxNA,funData-method}
\title{approxNA for funData objects}
\usage{
\S4method{approxNA}{funData}(object)
}
\description{
approxNA for funData objects
}
\keyword{internal}
