% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\docType{methods}
\name{argvals,irregFunData-method}
\alias{argvals,irregFunData-method}
\title{Get argvals slot for irregular functional data objects}
\usage{
\S4method{argvals}{irregFunData}(object)
}
\description{
Get argvals slot for irregular functional data objects
}
\seealso{
\code{\link{argvals}}
}
\keyword{internal}
