% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\docType{methods}
\name{argvals<-,funData-method}
\alias{argvals<-,funData-method}
\title{Set argvals slot for funData objects}
\usage{
\S4method{argvals}{funData}(object) <- value
}
\description{
Set argvals slot for funData objects
}
\seealso{
\code{\link{argvals<-}}
}
\keyword{internal}
