% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\docType{methods}
\name{argvals<-,irregFunData-method}
\alias{argvals<-,irregFunData-method}
\title{Set argvals slot for irregular functional objects}
\usage{
\S4method{argvals}{irregFunData}(object) <- value
}
\description{
Set argvals slot for irregular functional objects
}
\seealso{
\code{\link{argvals<-}}
}
\keyword{internal}
