% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\docType{methods}
\name{argvals<-,multiFunData-method}
\alias{argvals<-,multiFunData-method}
\title{Set argvals slot for multiFunData objects}
\usage{
\S4method{argvals}{multiFunData}(object) <- value
}
\description{
Set argvals slot for multiFunData objects
}
\seealso{
\code{\link{argvals<-}}
}
\keyword{internal}
