% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{Arith.funData}
\alias{Arith,funData,funData-method}
\alias{Arith,funData,numeric-method}
\alias{Arith,multiFunData,multiFunData-method}
\alias{Arith,multiFunData,numeric-method}
\alias{Arith,numeric,funData-method}
\alias{Arith,numeric,multiFunData-method}
\alias{Arith.funData}
\title{Arithmetics of functional data objects}
\usage{
\S4method{Arith}{funData,funData}(e1, e2)

\S4method{Arith}{funData,numeric}(e1, e2)

\S4method{Arith}{numeric,funData}(e1, e2)

\S4method{Arith}{multiFunData,multiFunData}(e1, e2)

\S4method{Arith}{multiFunData,numeric}(e1, e2)

\S4method{Arith}{numeric,multiFunData}(e1, e2)
}
\arguments{
\item{e1,e2}{Objects of class \code{funData}, \code{multiFunData} or
  \code{numeric}. If two functional data objects are used, they must be of
  the same class, have the same domain and the same number of observations.}
}
\value{
An object of the same functional data class as \code{e1} or
  \code{e2}, respectively.
}
\description{
These functions allow basic arithmetics for functional data and numerics
based on \code{\link[methods]{Arith}}.  The operations are made pointwise for
each observation.
}
\section{Warning}{
 Note that not all combinations of operations and classes
  make sense, e.g. \code{e1 ^ e2} is sensible if \code{e1} is of class
  \code{funData} or \code{multiFunData} and \code{e2} is numeric. The reverse
  is not true.
}
\examples{
oldpar <- par(no.readonly = TRUE)
par(mfrow = c(3,2), mar = rep(2.1,4))

xVal <- seq(0, 2*pi, 0.01)
object1 <- funData(xVal, outer(seq(0.75, 1.25, by = 0.05), sin(xVal)))
object2 <- funData(xVal, outer(seq(0.75, 1.25, by = 0.05), cos(xVal)))

plot(object1, main = "Object1")
plot(object2, main = "Object2")

# Only Functional data objects
plot(object1 + object2, main = "Sum")
plot(object1 - object2, main = "Difference")

# Mixed
plot(4*object1 + 5,  main = "4*Object1 + 5") # Note y-axis!
plot(object1^2 + object2^2, main = "Pythagoras")

par(oldpar)
}
\seealso{
\linkS4class{funData}, \linkS4class{multiFunData},
  \link[methods]{Arith}
}

