% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataClass.R
\docType{methods}
\name{as.funData}
\alias{as.funData}
\alias{as.funData,irregFunData-method}
\title{Coerce an irregFunData object to class funData}
\usage{
as.funData(object)

\S4method{as.funData}{irregFunData}(object)
}
\arguments{
\item{object}{The \code{irregFunData} object that is to be converted to a
  \code{funData} object with missing values.}
}
\description{
Coerce an \code{irregFunData} object to class \code{funData}.
}
\seealso{
\linkS4class{funData}, \linkS4class{irregFunData}
}

