% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataClass.R
\docType{methods}
\name{as.irregFunData}
\alias{as.irregFunData}
\alias{as.irregFunData,funData-method}
\title{Coerce a funData object to class irregFunData}
\usage{
as.irregFunData(object)

\S4method{as.irregFunData}{funData}(object)
}
\arguments{
\item{object}{The \code{funData} object that is to be converted to a
\code{irregFunData} object.}
}
\description{
This function coerces an object of class \code{funData} to a
\code{irregFunData} object.
}
\seealso{
\code{\linkS4class{funData}}, \code{\linkS4class{irregFunData}}
}

