% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{autoplot}
\alias{autoplot}
\alias{autolayer}
\title{ggplot Graphics for Functional Data Objects}
\usage{
autoplot(data, ...)

autolayer(data, ...)
}
\arguments{
\item{data}{A \code{funData}, \code{multiFunData} or
\code{irregFunData} object.}

\item{...}{Further parameters passed to the class-specific methods.}
}
\value{
A \code{\link[ggplot2]{ggplot} object}
}
\description{
The \pkg{funData} package allows to plot \code{funData},
\code{multiFunData} and \code{irregFunData} objects via \code{autoplot}
/ \code{autolayer} and the \pkg{ggplot2} package. For Details, see the
help pages of the class-specific plot functions (see below).
}
\seealso{
\code{\link{autoplot.funData}}/\code{\link{autolayer.funData}}
  for \code{funData} objects, \code{\link{autoplot.multiFunData}} for
  \code{multiFunData} objects and
  \code{\link{autoplot.irregFunData}}/\code{\link{autolayer.irregFunData}}
  for \code{irregFunData} objects.
}
