% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{dimSupp,funData-method}
\alias{dimSupp,funData-method}
\title{dimSupp for funData objects}
\usage{
\S4method{dimSupp}{funData}(object)
}
\description{
dimSupp for funData objects
}
\keyword{internal}

