% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{dimSupp,irregFunData-method}
\alias{dimSupp,irregFunData-method}
\title{dimSupp for irregular functional data objects}
\usage{
\S4method{dimSupp}{irregFunData}(object)
}
\description{
dimSupp for irregular functional data objects
}
\keyword{internal}
