% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{.intWeights}
\alias{.intWeights}
\title{Calculate weights for numerical integration}
\usage{
.intWeights(argvals, method = "trapezoidal")
}
\arguments{
\item{argvals}{A numeric vector of x-Values}

\item{method}{A character string, giving the numerical integration method to use (default is \code{trapezoidal}, alternatively use \code{midpoint})}
}
\value{
A vector of integration weights
}
\description{
This function calculates the weights for numerical integration
}
\seealso{
\code{\link{integrate}}
}
