% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{.scalarProduct}
\alias{.scalarProduct}
\title{Generic method for scalar products, based on integrate}
\usage{
.scalarProduct(object1, object2, ...)
}
\arguments{
\item{object1, object2}{Generic objects}

\item{...}{Further objects passed to \code{\link{integrate}}}
}
\description{
Generic method for scalar products, based on integrate
}
