% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{extractObs,funData-method}
\alias{extractObs,funData-method}
\title{extractObs for funData objects}
\usage{
\S4method{extractObs}{funData}(object, obs = 1:nObs(object),
  xVal = getxVal(object))
}
\description{
extractObs for funData objects
}
\keyword{internal}

