% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{extractObs,multiFunData-method}
\alias{extractObs,multiFunData-method}
\title{extractObs for multiFunData objects}
\usage{
\S4method{extractObs}{multiFunData}(object, obs = 1:nObs(object),
  xVal = getxVal(object))
}
\description{
extractObs for multiFunData objects
}
\keyword{internal}

