% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{extrapolateIrreg}
\alias{extrapolateIrreg}
\title{Extrapolate irregular functional data to a given domain}
\usage{
extrapolateIrreg(object, rangex = range(object@argvals))
}
\description{
This function extrapolates an \code{irregFunData} object to a given domain.
If only one point is observed, the function is extrapolated as a constant; in
all other cases it is extrapolated linearly.
}
\keyword{internal}

