% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{flipFuns,funData,irregFunData-method}
\alias{flipFuns,funData,irregFunData-method}
\title{Flip irregular functional data - funData as reference}
\usage{
\S4method{flipFuns}{funData,irregFunData}(refObject, newObject, ...)
}
\description{
Flip irregular functional data - funData as reference
}
\seealso{
\code{\link{flipFuns}}
}
\keyword{internal}

