% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{flipFuns,multiFunData,multiFunData-method}
\alias{flipFuns,multiFunData,multiFunData-method}
\title{Flip multivariate functional data}
\usage{
\S4method{flipFuns}{multiFunData,multiFunData}(refObject, newObject)
}
\description{
Flip multivariate functional data
}
\seealso{
\code{\link{flipFuns}}
}
\keyword{internal}

