% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataClass.R
\docType{methods}
\name{funData-constructor}
\alias{funData-constructor}
\title{Constructor for functional data objects, first argument (argvals) passed as list or vector of numerics}
\usage{
funData(argvals, X)
}
\arguments{
\item{argvals}{A list of numeric vectors or a single numeric vector, giving the 
sampling points in the domains. See Details.}

\item{X}{An array of dimension  \eqn{N \times M}{N x M} (for one-dimensional
domains, or \eqn{N \times M_1 \times \ldots \times M_d}{N x M_1 x \ldots x
M_d} for higher-dimensional domains), giving the observed values for
\eqn{N}{N} individuals. Missing values can be included via \code{NA}. See
Details.}
}
\description{
Constructor for functional data objects, first argument (argvals) passed as list or vector of numerics
}
\seealso{
\code{\linkS4class{funData}}
}
\keyword{internal}

