% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataClass.R
\name{funData-setAs}
\alias{funData-setAs}
\title{Coerce a funData object to class multiFunData}
\description{
Coerce a funData object to class multiFunData
}
\seealso{
\linkS4class{funData}, \link{as.multiFunData}
}
\keyword{internal}

