% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{getArgvals,funData-method}
\alias{getArgvals,funData-method}
\title{Get argvals slot for funData objects}
\usage{
\S4method{getArgvals}{funData}(object)
}
\description{
Get argvals slot for funData objects
}
\seealso{
\link{getArgvals}
}
\keyword{internal}

