% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{getArgvals,irregFunData-method}
\alias{getArgvals,irregFunData-method}
\title{Get argvals slot for irregular functional data objects}
\usage{
\S4method{getArgvals}{irregFunData}(object)
}
\description{
Get argvals slot for irregular functional data objects
}
\seealso{
\code{\link{getArgvals}}
}
\keyword{internal}
