% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{getArgvals,multiFunData-method}
\alias{getArgvals,multiFunData-method}
\title{Get argvals slot for multiFunData objects}
\usage{
\S4method{getArgvals}{multiFunData}(object)
}
\description{
Get argvals slot for multiFunData objects
}
\seealso{
\code{\link{getArgvals}}
}
\keyword{internal}

