% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{getX,irregFunData-method}
\alias{getX,irregFunData-method}
\title{Get X slot for irregular functional data objects}
\usage{
\S4method{getX}{irregFunData}(object)
}
\description{
Get X slot for irregular functional data objects
}
\seealso{
\code{\link{getX}}
}
\keyword{internal}

