% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{getX,multiFunData-method}
\alias{getX,multiFunData-method}
\title{Get X slot for multiFunData objects}
\usage{
\S4method{getX}{multiFunData}(object)
}
\description{
Get X slot for multiFunData objects
}
\seealso{
\code{\link{getX}}
}
\keyword{internal}

