% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{getxVal,funData-method}
\alias{getxVal,funData-method}
\title{Get xVal slot for funData objects}
\usage{
\S4method{getxVal}{funData}(object)
}
\description{
Get xVal slot for funData objects
}
\seealso{
\link{getxVal}
}
\keyword{internal}

