% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{getxVal,irregFunData-method}
\alias{getxVal,irregFunData-method}
\title{Get xVal slot for irregular functional data objects}
\usage{
\S4method{getxVal}{irregFunData}(object)
}
\description{
Get xVal slot for irregular functional data objects
}
\seealso{
\link{getxVal}
}
\keyword{internal}

