% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{getxVal,multiFunData-method}
\alias{getxVal,multiFunData-method}
\title{Get xVal slot for multiFunData objects}
\usage{
\S4method{getxVal}{multiFunData}(object)
}
\description{
Get xVal slot for multiFunData objects
}
\seealso{
\link{getxVal}
}
\keyword{internal}

