% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{ggplot}
\alias{ggplot}
\title{ggplot Graphics for Functional Data Objects}
\usage{
ggplot(data, ...)
}
\arguments{
\item{data}{A \code{funData}, \code{multiFunData} or \code{irregFunData} object.}

\item{...}{Further parameters passed to the class-specific methods.}
}
\value{
A \code{\link[ggplot2]{ggplot} object}
}
\description{
The \code{funData} allows to plot \code{funData}, \code{multiFunData} and \code{irregFunData} 
objects via the \pkg{ggplot2} package. For Details, see the help pages of the class-specific plot
functions (see below).
}
\seealso{
\code{\link{ggplot.funData}} for \code{funData} objects,
  \code{\link{ggplot.multiFunData}} for \code{multiFunData} objects and
  \code{\link{ggplot.irregFunData}} for \code{irregFunData} objects.
}
