% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{integrate,funData-method}
\alias{integrate,funData-method}
\title{Integrate method for funData objects}
\usage{
\S4method{integrate}{funData}(object, method = "trapezoidal")
}
\description{
Integrate method for funData objects
}
\seealso{
\code{\link{integrate}}, \code{\linkS4class{funData}}
}
\keyword{internal}
