% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{integrate,irregFunData-method}
\alias{integrate,irregFunData-method}
\title{Integrate method for irregular functional data objects}
\usage{
\S4method{integrate}{irregFunData}(object, method = "trapezoidal",
  fullDom = FALSE)
}
\description{
Integrate method for irregular functional data objects
}
\seealso{
\code{\link{integrate}}, \code{\linkS4class{irregFunData}}
}
\keyword{internal}

