% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{integrate,multiFunData-method}
\alias{integrate,multiFunData-method}
\title{Integrate method for multiFunData objects}
\usage{
\S4method{integrate}{multiFunData}(object, ...)
}
\description{
Integrate method for multiFunData objects
}
\seealso{
\code{\link{integrate}}, \code{\linkS4class{multiFunData}}
}
\keyword{internal}
