% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{integrate}
\alias{integrate}
\title{Integrate functional data}
\usage{
integrate(object, ...)
}
\arguments{
\item{object}{An object of class \code{funData} or \code{multiFunData}.}

\item{...}{Methods for internal function \code{.intWeights}, defaults to trapezoidal (alternative: midpoint).}
}
\value{
A vector of numerics, containing the integral values for each observation.
}
\description{
Integrate all observations of a \code{funData} object or a \code{multiFunData} object over their domain.
}
\details{
The function is currently implemented only for functional data with one- and two-dimensional domains.
}
\examples{
# Univariate
object <- funData(xVal = 1:5, X = rbind(1:5, 6:10))
integrate(object)

# Multivariate
multiObject <- multiFunData(object, funData(xVal = 1:3, X = rbind(3:5, 6:8)))
integrate(multiObject)
}
\seealso{
\linkS4class{funData}, \linkS4class{multiFunData}
}

