% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{integrate3D}
\alias{integrate3D}
\title{Integrate a function on a rectangular 3D grid}
\usage{
integrate3D(f, argvals)
}
\arguments{
\item{f}{A 3D array, representing the function evaluations on the grid}

\item{argvals}{A list with 3 elements, representing the grid points in the first, second and third dimension of f}
}
\value{
The result of the numerical integration of f on the given grid.
}
\description{
Integrate a function on a rectangular 3D grid
}
\keyword{internal}

