% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataClass.R
\docType{methods}
\name{irregFunData-constructor}
\alias{irregFunData-constructor}
\title{Constructor for irregular functional data objects}
\usage{
irregFunData(xVal, X)
}
\description{
Constructor for irregular functional data objects
}
\seealso{
\linkS4class{irregFunData}
}
\keyword{internal}

