% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{nObs,funData-method}
\alias{nObs,funData-method}
\title{nObs for funData objects}
\usage{
\S4method{nObs}{funData}(object)
}
\description{
nObs for funData objects
}
\keyword{internal}

