% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{nObs,irregFunData-method}
\alias{nObs,irregFunData-method}
\title{nObs for irregular functional data objects}
\usage{
\S4method{nObs}{irregFunData}(object)
}
\description{
nObs for irregular functional data objects
}
\keyword{internal}

