% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{nObs,multiFunData-method}
\alias{nObs,multiFunData-method}
\title{nObs for multiFunData objects}
\usage{
\S4method{nObs}{multiFunData}(object)
}
\description{
nObs for multiFunData objects
}
\keyword{internal}
