% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{nObs}
\alias{nObs}
\title{Get the number of observations}
\usage{
nObs(object)
}
\arguments{
\item{object}{An object of class \code{funData} or \code{multiFunData}.}
}
\value{
The number of observations in \code{object}.
}
\description{
This functions returns the number of observations in a \code{funData} or \code{multiFunData} object.
}
\examples{
# Univariate
object <- funData(xVal = 1:5, X = rbind(1:5, 6:10))
nObs(object)

# Multivariate
multiObject <- multiFunData(object, funData(xVal = 1:3, X = rbind(3:5, 6:8)))
nObs(multiObject)
}
\seealso{
\linkS4class{funData}, \linkS4class{multiFunData}
}

