% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{nObsPoints,funData-method}
\alias{nObsPoints,funData-method}
\title{nObsPoints for funData objects}
\usage{
\S4method{nObsPoints}{funData}(object)
}
\description{
nObsPoints for funData objects
}
\keyword{internal}

