% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{nObsPoints,irregFunData-method}
\alias{nObsPoints,irregFunData-method}
\title{nObsPoints for irregular functional data objects}
\usage{
\S4method{nObsPoints}{irregFunData}(object)
}
\description{
nObsPoints for irregular functional data objects
}
\keyword{internal}

