% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{nObsPoints}
\alias{nObsPoints}
\title{Get the number of observation points for irregular functional data}
\usage{
nObsPoints(object)
}
\arguments{
\item{object}{An object of class \code{irregFunData}.}
}
\value{
The number of observation points in each functions of \code{object}.
}
\description{
This functions returns the number of observation points in an
\code{irregFunData}  object.
}
\section{Warning}{
 Do not confound with \code{\link{nObs}}, which returns the
  number of observations (i.e. the number of observed functions) in an object
  of a functional data class.
}
\examples{
# Univariate (irregular)
irregObject <- irregFunData(xVal = list(1:5, 2:4), X = list(2:6, 3:5))
nObsPoints(irregObject)
}
\seealso{
\linkS4class{irregFunData}, \code{\link{extractObs}}
}

