% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{norm,funData-method}
\alias{norm,funData-method}
\title{Calculate the norm for univariate functional data}
\usage{
\S4method{norm}{funData}(object, squared = TRUE, obs = 1:nObs(object),
  method = "trapezoidal")
}
\description{
Calculate the norm for univariate functional data
}
\seealso{
\link{norm} \linkS4class{funData}
}
\keyword{internal}

