% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{norm,funData,missing-method}
\alias{norm,funData,missing-method}
\title{Calculate the norm for univariate functional data}
\usage{
\S4method{norm}{funData,missing}(x, squared = TRUE, obs = 1:nObs(x),
  method = "trapezoidal", weight = 1)
}
\description{
Calculate the norm for univariate functional data
}
\seealso{
\code{\link{norm}}, \code{\linkS4class{funData}}
}
\keyword{internal}

