% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{norm,irregFunData-method}
\alias{norm,irregFunData-method}
\title{Calculate the norm for irregular functional data}
\usage{
\S4method{norm}{irregFunData}(object, squared = TRUE, obs = 1:nObs(object),
  method = "trapezoidal", weight = 1, fullDom = FALSE)
}
\description{
Calculate the norm for irregular functional data
}
\seealso{
\code{\link{norm}} \code{\linkS4class{irregFunData}}
}
\keyword{internal}

