% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{norm,multiFunData-method}
\alias{norm,multiFunData-method}
\title{Calculate the norm for multivariate functional data}
\usage{
\S4method{norm}{multiFunData}(object, squared = TRUE, obs = 1:nObs(object),
  method = "trapezoidal")
}
\description{
Calculate the norm for multivariate functional data
}
\seealso{
\link{norm} \linkS4class{multiFunData}
}
\keyword{internal}

