% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{norm,multiFunData,missing-method}
\alias{norm,multiFunData,missing-method}
\title{Calculate the norm for multivariate functional data}
\usage{
\S4method{norm}{multiFunData,missing}(
  x,
  squared = TRUE,
  obs = seq_len(nObs(x)),
  method = "trapezoidal",
  weight = rep(1, length(x))
)
}
\description{
Calculate the norm for multivariate functional data
}
\seealso{
\code{\link{norm}}, \code{\linkS4class{multiFunData}}
}
\keyword{internal}
