% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{norm.funData}
\alias{norm.funData}
\title{Calculate the norm for univariate functional data}
\usage{
norm.funData(object, squared, obs, method, weight)
}
\description{
Calculate the norm for univariate functional data
}
\seealso{
\link{norm}
}
\keyword{internal}

