% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{norm.irregFunData}
\alias{norm.irregFunData}
\title{Calculate the norm for irregular functional data}
\usage{
norm.irregFunData(object, squared, obs, method, weight, fullDom)
}
\description{
Calculate the norm for irregular functional data
}
\seealso{
\code{\link{norm}}
}
\keyword{internal}

