% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{norm}
\alias{norm}
\title{Calculate the norm of functional data}
\usage{
norm(object, ...)
}
\arguments{
\item{object}{An object of class \code{funData} or \code{multiFunData}.}

\item{...}{Further parameters (see Details).}
}
\value{
A numeric vector representing the norm of each observation.
}
\description{
This function calculates the norm for each observation of a \code{funData} or
\code{multiFunData} object.
}
\details{
Further parameters passed to this function may include:
\itemize{
  \item \code{squared}: Logical. If \code{TRUE} (default), the function calculates the
  squared norm, otherwise the result is not squared.
  \item \code{obs}: A numeric vector, giving the indices of the observations, for
  which the norm is to be calculated. Defaults to all observations.
  \item \code{method}: A character string, giving the integration method to be used. See \link{integrate} for details.
}
}
\section{Warning}{

 The function is currently implemented only for functional data with one- and
two-dimensional domains.
}
\examples{
# Univariate
object <- funData(xVal = 1:5, X = rbind(1:5, 6:10))
norm(object)

# Multivariate
multiObject <- multiFunData(object, funData(xVal = 1:3, X = rbind(3:5, 6:8)))
norm(multiObject)
}
\seealso{
\linkS4class{funData}, \linkS4class{multiFunData}, \code{\link{integrate}}
}

