% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{plot.multiFunData}
\alias{plot,multiFunData,missing-method}
\alias{plot.multiFunData}
\title{Plotting multivariate functional data}
\usage{
plot.multiFunData(x, y, obs = 1:nObs(x), dim = 1:length(x),
  par.plot = NULL, add = FALSE, main = NULL, xlab = "argvals",
  ylab = "", ...)

\S4method{plot}{multiFunData,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{multiFunData}.}

\item{y}{Missing.}

\item{obs}{A vector of numerics giving the observations to plot. Defaults to
all observations in \code{x}. For two-dimensional functions (images)
\code{obs} must have length 1.}

\item{dim}{The dimensions to plot. Defaults \code{length(x)}.}

\item{par.plot}{Graphic parameters to be passed to the plotting regions. The
option \code{mfrow} is ignored. Defaults to \code{FALSE}. See
\code{\link[graphics]{par}} for details.}

\item{add}{Logical. If \code{TRUE}, add to current plot (only for
one-dimensional functions). Defaults to \code{FALSE}.}

\item{main}{A string vector, giving the title of the plot. Can have the same
length as \code{dim} (different titles for each dimension) or length
\code{1} (one title for all dimensions). Defaults to \code{NULL}.}

\item{xlab,ylab}{The titles for x- and y-axis. Defaults to "argvals" for the
x-axis and no title for the y-axis for all elements. Can be supplied as a
vector of the same length as \code{x} (one x-/y-lab for each element) or a
single strin that is applied for all elements. See
\code{\link[graphics]{plot}} for details.}

\item{...}{Additional arguments to \code{plot}.}
}
\description{
This function plots observations of multivariate functional data on their
domain. The graphic device is split in a number of subplots (specified by
\code{dim}) via \code{mfrow} (\code{\link[graphics]{par}}) and the univariate
elements are plotted using \code{plot}.
}
\section{Warning}{
 The function is currently implemented only for functional
  data with one- and two-dimensional domains.
}
\examples{
oldpar <- par(no.readonly = TRUE)
argvals <- seq(0, 2*pi, 0.1)

# One-dimensional elements
object1 <- funData(argvals, outer(seq(0.75, 1.25, length.out = 11), sin(argvals)))
object2 <- funData(argvals, outer(seq(0.75, 1.25, length.out = 11), cos(argvals)))

multiObject <- multiFunData(object1, object2)
plot(multiObject, main = c("1st element", "2nd element")) # different titles
plot(multiObject, main = "Multivariate Functional Data") # one title for all

# Mixed-dimensional elements
X <- array(0, dim = c(11, length(argvals), length(argvals)))
X[1,,] <- outer(argvals, argvals, function(x,y){sin((x-pi)^2 + (y-pi)^2)})
object2D <- funData(list(argvals, argvals), X)

multiObject <- multiFunData(object1, object2D)
# different titles and labels
plot(multiObject, main = c("1st element", "2nd element"), obs = 1,
     xlab = c("xlab1", "xlab2"),
     ylab = "one ylab for all")
# one title for all
plot(multiObject, main = "Multivariate Functional Data", obs = 1)

\dontrun{plot(multiObject, main = c("1st element", "2nd element")) # must specify obs!}
\dontrun{plot(multiObject, main = "Multivariate Functional Data") # must specify obs!}

par(oldpar)
}
\seealso{
\linkS4class{funData}, \linkS4class{multiFunData},
  \code{\link{plot.funData}}
}

