% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{print.funData}
\alias{print.funData}
\title{A print method for univariate functional data}
\usage{
\method{print}{funData}(x, ...)
}
\arguments{
\item{x}{A \code{funData} object.}
}
\description{
This function prints basic information about a \code{funData} object. This is
the standard console output for \code{funData} objects.
}
\keyword{internal}
