% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{print.irregFunData}
\alias{print.irregFunData}
\title{A print method for irregular functional data}
\usage{
\method{print}{irregFunData}(x, ...)
}
\arguments{
\item{x}{An \code{irregFunData} object.}
}
\description{
This function prints basic information about a \code{irregFunData} object. This is
the standard console output for \code{irregFunData} objects.
}
\keyword{internal}

