% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{scalarProduct,funData,funData-method}
\alias{scalarProduct,funData,funData-method}
\title{Scalar product for functional data}
\usage{
\S4method{scalarProduct}{funData,funData}(object1, object2, ...)
}
\description{
Scalar product for functional data
}
\seealso{
\code{\link{integrate}}, \code{\link{norm}}.
}
\keyword{internal}
