% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{scalarProduct,irregFunData,irregFunData-method}
\alias{scalarProduct,irregFunData,irregFunData-method}
\title{Scalar product for irregular functional data}
\usage{
\S4method{scalarProduct}{irregFunData,irregFunData}(object1, object2, ...)
}
\description{
Scalar product for irregular functional data
}
\seealso{
\code{\link{integrate}}, \code{\link{norm}}.
}
\keyword{internal}
