% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{scalarProduct,multiFunData,multiFunData-method}
\alias{scalarProduct,multiFunData,multiFunData-method}
\title{Scalar product for multivariate functional data}
\usage{
\S4method{scalarProduct}{multiFunData,multiFunData}(object1, object2,
  weight = rep(1, length(object1)), ...)
}
\description{
Scalar product for multivariate functional data
}
\seealso{
\code{\link{integrate}}, \code{\link{norm}}.
}
\keyword{internal}

