% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{setArgvals,funData-method}
\alias{setArgvals,funData-method}
\title{Set argvals slot for funData objects}
\usage{
\S4method{setArgvals}{funData}(object, newArgvals)
}
\description{
Set argvals slot for funData objects
}
\seealso{
\code{\link{setArgvals}}
}
\keyword{internal}
