% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{setArgvals,irregFunData-method}
\alias{setArgvals,irregFunData-method}
\title{Set argvals slot for irregular functional objects}
\usage{
\S4method{setArgvals}{irregFunData}(object, newArgvals)
}
\description{
Set argvals slot for irregular functional objects
}
\seealso{
\code{\link{setArgvals}}
}
\keyword{internal}
