% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{setArgvals,multiFunData-method}
\alias{setArgvals,multiFunData-method}
\title{Set argvals slot for multiFunData objects}
\usage{
\S4method{setArgvals}{multiFunData}(object, newArgvals)
}
\description{
Set argvals slot for multiFunData objects
}
\seealso{
\code{\link{setArgvals}}
}
\keyword{internal}

