% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{setX,funData-method}
\alias{setX,funData-method}
\title{Set X slot for funData objects}
\usage{
\S4method{setX}{funData}(object, newX)
}
\description{
Set X slot for funData objects
}
\seealso{
\code{\link{setX}}
}
\keyword{internal}
