% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{setX,irregFunData-method}
\alias{setX,irregFunData-method}
\title{Set X slot for irregular functional data objects}
\usage{
\S4method{setX}{irregFunData}(object, newX)
}
\description{
Set X slot for irregular functional data objects
}
\seealso{
\code{\link{setX}}
}
\keyword{internal}

