% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{setX,multiFunData-method}
\alias{setX,multiFunData-method}
\title{Set X slot for multiFunData objects}
\usage{
\S4method{setX}{multiFunData}(object, newX)
}
\description{
Set X slot for multiFunData objects
}
\seealso{
\code{\link{setX}}
}
\keyword{internal}
