% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{setxVal,funData-method}
\alias{setxVal,funData-method}
\title{Set xVal slot for funData objects}
\usage{
\S4method{setxVal}{funData}(object, newxVal)
}
\description{
Set xVal slot for funData objects
}
\seealso{
\link{setxVal}
}
\keyword{internal}

