% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{setxVal,multiFunData-method}
\alias{setxVal,multiFunData-method}
\title{Set xVal slot for multiFunData objects}
\usage{
\S4method{setxVal}{multiFunData}(object, newxVal)
}
\description{
Set xVal slot for multiFunData objects
}
\seealso{
\link{setxVal}
}
\keyword{internal}

