% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{tensorProduct,funData-method}
\alias{tensorProduct,funData-method}
\title{Tensor product for functional data}
\usage{
\S4method{tensorProduct}{funData}(...)
}
\description{
Tensor product for functional data
}
\seealso{
\code{\link{tensorProduct}}
}
\keyword{internal}
