% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\name{X,funData-method}
\alias{X,funData-method}
\title{Get X slot for funData objects}
\usage{
\S4method{X}{funData}(object)
}
\description{
Get X slot for funData objects
}
\seealso{
\code{\link{X}}
}
\keyword{internal}
