% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\docType{methods}
\name{X,irregFunData-method}
\alias{X,irregFunData-method}
\title{Get X slot for irregular functional data objects}
\usage{
\S4method{X}{irregFunData}(object)
}
\description{
Get X slot for irregular functional data objects
}
\seealso{
\code{\link{X}}
}
\keyword{internal}
