% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\docType{methods}
\name{X,multiFunData-method}
\alias{X,multiFunData-method}
\title{Get X slot for multiFunData objects}
\usage{
\S4method{X}{multiFunData}(object)
}
\description{
Get X slot for multiFunData objects
}
\seealso{
\code{\link{X}}
}
\keyword{internal}
