% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\docType{methods}
\name{X<-,funData-method}
\alias{X<-,funData-method}
\title{Set X slot for funData objects}
\usage{
\S4method{X}{funData}(object) <- value
}
\description{
Set X slot for funData objects
}
\keyword{internal}
