% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\name{X<-,multiFunData-method}
\alias{X<-,multiFunData-method}
\title{Set X slot for multiFunData objects}
\usage{
\S4method{X}{multiFunData}(object) <- value
}
\description{
Set X slot for multiFunData objects
}
\keyword{internal}
