% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataClass.R
\docType{methods}
\name{as.funData}
\alias{as.funData}
\alias{as.funData,irregFunData-method}
\title{Coerce an irregFunData object to class funData}
\usage{
as.funData(object)

\S4method{as.funData}{irregFunData}(object)
}
\arguments{
\item{object}{The \code{irregFunData} object that is to be converted to a
\code{funData} object with missing values.}
}
\description{
This function coerces an object of class \code{irregFunData} to a
\code{funData} object with missing values, which is defined on the union of
all observation points.
}
\seealso{
\code{\linkS4class{funData}}, \code{\linkS4class{irregFunData}}
}

