% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataClass.R
\docType{methods}
\name{as.multiFunData}
\alias{as.multiFunData}
\alias{as.multiFunData,funData-method}
\title{Coerce a funData object to class multiFunData}
\usage{
as.multiFunData(object)

\S4method{as.multiFunData}{funData}(object)
}
\arguments{
\item{object}{The \code{funData} object that is to be converted to a
  \code{multiFunData} object of length 1.}
}
\description{
Coerce a \code{funData} object to class \code{multiFunData} with one element.
}
\seealso{
\linkS4class{funData}, \linkS4class{multiFunData}
}

