% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{dimSupp,multiFunData-method}
\alias{dimSupp,multiFunData-method}
\title{dimSupp for multiFunData objects}
\usage{
\S4method{dimSupp}{multiFunData}(object)
}
\description{
dimSupp for multiFunData objects
}
\keyword{internal}

