% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{extractObs,funData-method}
\alias{extractObs,funData-method}
\title{extractObs for funData objects}
\usage{
\S4method{extractObs}{funData}(object, obs = seq_len(nObs(object)),
  argvals = funData::argvals(object))
}
\description{
extractObs for funData objects
}
\keyword{internal}
