% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{extractObs,irregFunData-method}
\alias{extractObs,irregFunData-method}
\title{extractObs for irregular functional data}
\usage{
\S4method{extractObs}{irregFunData}(
  object,
  obs = seq_len(nObs(object)),
  argvals = funData::argvals(object)
)
}
\description{
extractObs for irregular functional data
}
\keyword{internal}
