% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{extractObs,irregFunData-method}
\alias{extractObs,irregFunData-method}
\title{extractObs for irregular functional data}
\usage{
\S4method{extractObs}{irregFunData}(object, obs = 1:nObs(object),
  xVal = getxVal(object))
}
\description{
extractObs for irregular functional data
}
\keyword{internal}

