% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{extractObs,multiFunData-method}
\alias{extractObs,multiFunData-method}
\title{extractObs for multiFunData objects}
\usage{
\S4method{extractObs}{multiFunData}(object, obs = 1:nObs(object),
  argvals = funData::argvals(object))
}
\description{
extractObs for multiFunData objects
}
\keyword{internal}
