% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{flipFuns,funData,funData-method}
\alias{flipFuns,funData,funData-method}
\title{Flip univariate functional data}
\usage{
\S4method{flipFuns}{funData,funData}(refObject, newObject)
}
\description{
Flip univariate functional data
}
\seealso{
\code{\link{flipFuns}}
}
\keyword{internal}
