% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataClass.R
\name{funData-setAs}
\alias{funData-setAs}
\title{Coerce a funData object to class multiFunData}
\description{
Coerce a funData object to class multiFunData

Coerce a funData object to class irregFunData
}
\examples{
# create funData object with 2 observations
f <- simFunData(N = 5, M = 7, eValType = "linear",
                eFunType = "Fourier", argvals = seq(0,1,0.01))$simData

# sparsify artificially
fSparse <- sparsify(f, minObs = 4, maxObs = 10)

# coerce to irregFunData object
i <- as.irregFunData(fSparse)
i
}
\seealso{
\code{\linkS4class{funData}}, \code{\link{as.multiFunData}}

\code{\linkS4class{funData}}, \code{\link{as.irregFunData}}
}
\keyword{internal}

