% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{getX,funData-method}
\alias{getX,funData-method}
\title{Get X slot for funData objects}
\usage{
\S4method{getX}{funData}(object)
}
\description{
Get X slot for funData objects
}
\seealso{
\link{getX}
}
\keyword{internal}

