% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMethods.R
\docType{methods}
\name{ggplot.irregFunData}
\alias{ggplot,irregFunData-method}
\alias{ggplot.irregFunData}
\title{Visualize irregular functional data objects using ggplot}
\usage{
ggplot.irregFunData(data, obs = 1:nObs(data), ...)

\S4method{ggplot}{irregFunData}(data, ...)
}
\arguments{
\item{data}{A \code{irregFunData} object.}

\item{obs}{A vector of numerics giving the observations to plot. Defaults to
all observations in \code{data}. For two-dimensional functions (images)
\code{obs} must have length 1.}

\item{...}{Further parameters passed to \code{\link[ggplot2]{geom_line}}, e.g. \code{alpha, color, fill, linetype, size}).}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be customized using
  all functionalities of the \pkg{ggplot2} package.
}
\description{
This function allows to plot \code{irregFunData} objects on their domain based on the
\pkg{ggplot2} package. The function provides a wrapper that rearranges the
data in a \code{irregFunData} object returns a basic \code{\link[ggplot2]{ggplot}} object, which can be
customized using all functionalities of the \pkg{ggplot2} package.
}
\examples{
# Generate data
argvals <- seq(0,2*pi,0.01)
ind <- replicate(5, sort(sample(1:length(argvals), sample(5:10,1))))
object <- irregFunData(argvals = lapply(ind, function(i){argvals[i]}),
                  X = lapply(ind, function(i){sample(1:10,1) / 10 * argvals[i]^2}))

# Plot the data
ggplot(object)

 # Parameters passed to geom_line are passed via the ... argument
ggplot(object, color = "red", linetype = 3)

# New layers can be added directly to the ggplot object
g <- ggplot(object)
g + theme_bw() + ggtitle("Plot with minimal theme and axis labels") +
 xlab("The x-Axis") + ylab("The y-Axis")
}
\seealso{
\code{\linkS4class{irregFunData}}, \code{\link[ggplot2]{ggplot}},
  \code{\link{plot.irregFunData}}
}

