% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{ggplot}
\alias{ggplot}
\alias{ggplot,funData-method}
\alias{ggplot,multiFunData-method}
\alias{ggplot,irregFunData-method}
\title{ggplot Graphics for Functional Data Objects}
\usage{
ggplot(data, ...)

\S4method{ggplot}{funData}(data, add = FALSE, ...)

\S4method{ggplot}{multiFunData}(data, ...)

\S4method{ggplot}{irregFunData}(data, add = FALSE, ...)
}
\arguments{
\item{data}{A \code{funData}, \code{multiFunData} or
\code{irregFunData} object.}

\item{...}{Further parameters passed to the class-specific methods.}

\item{add}{Logical. If \code{TRUE}, add to current plot (only for 
one-dimensional functions). Defaults to \code{FALSE}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
This function is deprecated. Use \code{\link{autoplot.funData}} /
\code{\link{autolayer.funData}} for \code{funData} objects,
\code{\link{autoplot.multiFunData}} for \code{multiFunData} objects and
\code{\link{autoplot.irregFunData}} /
\code{\link{autolayer.irregFunData}} for \code{irregFunData} objects
instead.
}
\details{
In the default case, this function calls \link[ggplot2]{ggplot} (if available).
}
\seealso{
\code{\link[ggplot2]{ggplot}},
  \code{\link[ggplot2]{autoplot}}, \code{\link[ggplot2]{autolayer}}
  from package \pkg{ggplot2}
}
