% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{integrate,funData-method}
\alias{integrate,funData-method}
\title{Integrate method for funData objects}
\usage{
\S4method{integrate}{funData}(object, method = "trapezoidal")
}
\description{
Integrate method for funData objects
}
\seealso{
\link{integrate} \linkS4class{funData}
}
\keyword{internal}

