% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataClass.R
\name{irregFunData-setAs}
\alias{irregFunData-setAs}
\title{Coerce an irregFunData object to class funData}
\description{
Coerce an irregFunData object to class funData
}
\seealso{
\linkS4class{funData}, \link{as.funData}
}
\keyword{internal}

