% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataClass.R
\name{irregFunData-setAs}
\alias{irregFunData-setAs}
\title{Coerce an irregFunData object to class funData}
\description{
Coerce an irregFunData object to class funData
}
\examples{
# create irregFunData object with 2 observations
i1 <- irregFunData(argvals = list(1:5, 3:6), X = list(2:6, 4:7))
i1@argvals # argvals and X are both lists with 2 entries
i1@X

# coerce to funData object (with missing values)
f1 <- as.funData(i1)
# argvals is a list containing one vector
# (one-dimensional domain, union of all observation points)
f1@argvals
# X is a matrix with 2 rows and missing values
f1@X
}
\seealso{
\code{\linkS4class{funData}}, \code{\link{as.funData}}
}
\keyword{internal}

