% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{meanFunction,funData-method}
\alias{meanFunction,funData-method}
\title{Mean for functional data}
\usage{
\S4method{meanFunction}{funData}(object, na.rm = FALSE)
}
\description{
Mean for functional data
}
\seealso{
\code{\link{meanFunction}}
}
\keyword{internal}

