% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{meanFunction,irregFunData-method}
\alias{meanFunction,irregFunData-method}
\title{Mean for irregular functional data}
\usage{
\S4method{meanFunction}{irregFunData}(object, na.rm = FALSE)
}
\description{
Mean for irregular functional data
}
\seealso{
\code{\link{meanFunction}}
}
\keyword{internal}
