% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{meanFunction,multiFunData-method}
\alias{meanFunction,multiFunData-method}
\title{Mean for multivariate functional data}
\usage{
\S4method{meanFunction}{multiFunData}(object, na.rm = FALSE)
}
\description{
Mean for multivariate functional data
}
\seealso{
\code{\link{meanFunction}}
}
\keyword{internal}
