% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataClass.R
\docType{methods}
\name{multiFunData-constructor}
\alias{multiFunData-constructor}
\title{Constructor for multivariate functional data objects}
\usage{
multiFunData(...)
}
\description{
Constructor for multivariate functional data objects
}
\seealso{
\code{\linkS4class{multiFunData}}
}
\keyword{internal}
