% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{nObs,multiFunData-method}
\alias{nObs,multiFunData-method}
\title{nObs for multiFunData objects}
\usage{
\S4method{nObs}{multiFunData}(object)
}
\description{
nObs for multiFunData objects
}
\keyword{internal}

