% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funDataMethods.R
\name{nObsPoints,multiFunData-method}
\alias{nObsPoints,multiFunData-method}
\title{nObsPoints for multiFunData objects}
\usage{
\S4method{nObsPoints}{multiFunData}(object)
}
\description{
nObsPoints for multiFunData objects
}
\keyword{internal}
