% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funDataMethods.R
\docType{methods}
\name{norm,multiFunData,missing-method}
\alias{norm,multiFunData,missing-method}
\title{Calculate the norm for multivariate functional data}
\usage{
\S4method{norm}{multiFunData,missing}(x, squared = TRUE, obs = 1:nObs(x),
  method = "trapezoidal", weight = rep(1, length(x)))
}
\description{
Calculate the norm for multivariate functional data
}
\seealso{
\code{\link{norm}}, \code{\linkS4class{multiFunData}}
}
\keyword{internal}

